// packages/worker/src/gtfs-proto.d.ts
export namespace transit_realtime {
// ...
    namespace FeedMessage {

        /** Properties of a FeedMessage. */
        interface $Properties {

            /** FeedMessage header */
            header: transit_realtime.FeedHeader.$Properties;

            /** FeedMessage entity */
            entity?: (transit_realtime.FeedEntity.$Properties[]|null);

            /** Unknown fields preserved while decoding when enabled */
            $unknowns?: Uint8Array[];
        }

        /** Shape of a FeedMessage. */
        type $Shape = transit_realtime.FeedMessage.$Properties;
    }
// ...
    class TripDescriptor {

        /**
         * Constructs a new TripDescriptor.
         * @param [properties] Properties to set
         */
        constructor(properties?: transit_realtime.TripDescriptor.$Properties);

        /** TripDescriptor .nyctTripDescriptor */
        ".nyctTripDescriptor"?: (NyctTripDescriptor.$Properties|null);
// ...
    }
// ...
}
// ...
export class NyctTripDescriptor {

    /**
     * Constructs a new NyctTripDescriptor.
     * @param [properties] Properties to set
     */
    constructor(properties?: NyctTripDescriptor.$Properties);

    /** Unknown fields preserved while decoding when enabled */
    $unknowns?: Uint8Array[];

    /** NyctTripDescriptor trainId. */
    trainId: string;

    /** NyctTripDescriptor isAssigned. */
    isAssigned: boolean;

    /** NyctTripDescriptor direction. */
    direction: NyctTripDescriptor.Direction;

    /**
     * Creates a new NyctTripDescriptor instance using the specified properties.
     * @param [properties] Properties to set
     * @returns NyctTripDescriptor instance
     */
    static create(properties: NyctTripDescriptor.$Shape): NyctTripDescriptor & NyctTripDescriptor.$Shape;
    static create(properties?: NyctTripDescriptor.$Properties): NyctTripDescriptor;

    /**
     * Encodes the specified NyctTripDescriptor message. Does not implicitly {@link NyctTripDescriptor.verify|verify} messages.
     * @param message NyctTripDescriptor message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    static encode(message: NyctTripDescriptor.$Properties, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Encodes the specified NyctTripDescriptor message, length delimited. Does not implicitly {@link NyctTripDescriptor.verify|verify} messages.
     * @param message NyctTripDescriptor message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    static encodeDelimited(message: NyctTripDescriptor.$Properties, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Decodes a NyctTripDescriptor message from the specified reader or buffer.
     * @param reader Reader or buffer to decode from
     * @param [length] Message length if known beforehand
     * @returns {NyctTripDescriptor & NyctTripDescriptor.$Shape} NyctTripDescriptor
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): NyctTripDescriptor & NyctTripDescriptor.$Shape;

    /**
     * Decodes a NyctTripDescriptor message from the specified reader or buffer, length delimited.
     * @param reader Reader or buffer to decode from
     * @returns {NyctTripDescriptor & NyctTripDescriptor.$Shape} NyctTripDescriptor
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    static decodeDelimited(reader: ($protobuf.Reader|Uint8Array)): NyctTripDescriptor & NyctTripDescriptor.$Shape;

    /**
     * Verifies a NyctTripDescriptor message.
     * @param message Plain object to verify
     * @returns `null` if valid, otherwise the reason why it is not
     */
    static verify(message: { [k: string]: any }): (string|null);

    /**
     * Creates a NyctTripDescriptor message from a plain object. Also converts values to their respective internal types.
     * @param object Plain object
     * @returns NyctTripDescriptor
     */
    static fromObject(object: { [k: string]: any }): NyctTripDescriptor;

    /**
     * Creates a plain object from a NyctTripDescriptor message. Also converts values to other types if specified.
     * @param message NyctTripDescriptor
     * @param [options] Conversion options
     * @returns Plain object
     */
    static toObject(message: NyctTripDescriptor, options?: $protobuf.IConversionOptions): { [k: string]: any };

    /**
     * Converts this NyctTripDescriptor to JSON.
     * @returns JSON object
     */
    toJSON(): { [k: string]: any };

    /**
     * Gets the type url for NyctTripDescriptor
     * @param [prefix] Custom type url prefix, defaults to `"type.googleapis.com"`
     * @returns The type url
     */
    static getTypeUrl(prefix?: string): string;
}

export namespace NyctTripDescriptor {

    /** Properties of a NyctTripDescriptor. */
    interface $Properties {

        /** NyctTripDescriptor trainId */
        trainId?: (string|null);

        /** NyctTripDescriptor isAssigned */
        isAssigned?: (boolean|null);

        /** NyctTripDescriptor direction */
        direction?: (NyctTripDescriptor.Direction|null);

        /** Unknown fields preserved while decoding when enabled */
        $unknowns?: Uint8Array[];
    }

    /** Shape of a NyctTripDescriptor. */
    type $Shape = NyctTripDescriptor.$Properties;

    /** Direction enum. */
    enum Direction {

        /** NORTH value */
        NORTH = 1,

        /** EAST value */
        EAST = 2,

        /** SOUTH value */
        SOUTH = 3,

        /** WEST value */
        WEST = 4
    }
}